MAINT: not quite functional yet, but close
