+++ mosfit/modules/objectives/likelihood.py
@@ -6,7 +6,6 @@ import scipy
 
 from mosfit.constants import LIKELIHOOD_FLOOR
 from mosfit.modules.module import Module
-from mosfit.utils import flux_density_unit
 
 
 # Important: Only define one ``Module`` class per file.
@@ -19,14 +18,14 @@ class Likelihood(Module):
 
     def process(self, **kwargs):
         """Process module."""
-        self.preprocess(**kwargs)
-        self._score_modifier = kwargs.get(self.key('score_modifier'), 0.0)
         self._fractions = kwargs['fractions']
         self._model_observations = kwargs['model_observations']
+        self._score_modifier = kwargs.get(self.key('score_modifier'), 0.0)
         self._upper_limits = np.array(kwargs.get('upperlimits', []),
                                       dtype=bool)
 
         ret = {'value': LIKELIHOOD_FLOOR}
+        value = ret['value']
 
         if min(self._fractions) < 0.0 or max(self._fractions) > 1.0:
             return ret
@@ -34,46 +33,13 @@ class Likelihood(Module):
             if not self._upper_limits[oi] and (isnan(obs) or
                                                not np.isfinite(obs)):
                 return ret
-        value = kwargs.get('value', LIKELIHOOD_FLOOR)
-        self._model_observations[self._cmask] = -2.5 * np.log10(
-            self._model_observations[self._cmask])
 
-        # Calculate (model - obs) residuals.
-        residuals = np.array([
-            (abs(x - ct) if not u or (x < ct and not isnan(x)) else 0.0)
-            if t == 'countrate' and ct is not None
-            else
-            (abs(x - y) if not u or (x < y and not isnan(x)) else 0.0)
-            if t == 'magnitude' and y is not None
-            else
-            (abs(x - fd) if not u or (x > fd and not isnan(x)) else 0.0)
-            if t == 'fluxdensity' and fd is not None else None
-            for x, y, ct, fd, u, o, t in zip(
-                self._model_observations, self._mags, self._cts, self._fds,
-                self._upper_limits, self._observed, self._observation_types)
-            if o
-        ])
+        diag = kwargs.get('kdiagonal', None)
+        residuals = kwargs.get('kresiduals', None)
 
-        if np.any(residuals == None):  # noqa: E711
-            raise ValueError('Null residual.')
+        if diag is None or residuals is None:
+            return ret
 
-        # Observational errors to be put in diagonal of error matrix.
-        diag = np.array([
-            ((ctel if x > ct else cteu) ** 2)
-            if t == 'countrate' and ct is not None else
-            ((el if x > y else eu) ** 2)
-            if t == 'magnitude' and y is not None else
-            ((fdel if x < fd else fdeu) ** 2)
-            if t == 'fluxdensity' and fd is not None else None
-            for x, y, eu, el, fd, fdeu, fdel, ct, ctel, cteu, o, t in zip(
-                self._model_observations, self._mags,
-                self._e_u_mags, self._e_l_mags, self._fds, self._e_u_fds,
-                self._e_l_fds, self._cts, self._e_l_cts, self._e_u_cts,
-                self._observed, self._observation_types) if o
-        ])
-
-        if np.any(diag == None):  # noqa: E711
-            raise ValueError('Null error.')
         if 'kmat' in kwargs:
             kmat = kwargs['kmat']
 
@@ -126,8 +92,10 @@ class Likelihood(Module):
                     np.matmul(np.matmul(residuals.T,
                                         scipy.linalg.inv(kmat)),
                               residuals) + np.log(scipy.linalg.det(kmat)))
-        elif 'ksmat' in kwargs:
-            raise RuntimeError('Should not have ksmat in likelihood!')
+            ret['kdiagonal'] = diag
+            ret['kresiduals'] = residuals
+        elif 'kfmat' in kwargs:
+            raise RuntimeError('Should not have kfmat in likelihood!')
         else:
             # Shortcut when matrix is diagonal.
             self._o_band_vs = kwargs['obandvs']
@@ -140,93 +108,3 @@ class Likelihood(Module):
             return ret
         ret['value'] = max(LIKELIHOOD_FLOOR, score)
         return ret
-
-    def preprocess(self, **kwargs):
-        """Construct arrays of observations based on data keys."""
-        self._mags = np.array(kwargs.get('magnitudes', []))
-        self._fds = np.array(kwargs.get('fluxdensities', []))
-        self._cts = np.array(kwargs.get('countrates', []))
-        self._freqs = kwargs.get('frequencies', [])
-        self._e_u_mags = kwargs.get('e_upper_magnitudes', [])
-        self._e_l_mags = kwargs.get('e_lower_magnitudes', [])
-        self._e_mags = kwargs.get('e_magnitudes', [])
-        self._e_u_fds = kwargs.get('e_upper_fluxdensities', [])
-        self._e_l_fds = kwargs.get('e_lower_fluxdensities', [])
-        self._e_fds = kwargs.get('e_fluxdensities', [])
-        self._u_fds = kwargs.get('u_fluxdensities', [])
-        self._u_freqs = kwargs.get('u_frequencies', [])
-        self._e_u_cts = kwargs.get('e_upper_countrates', [])
-        self._e_l_cts = kwargs.get('e_lower_countrates', [])
-        self._e_cts = kwargs.get('e_countrates', [])
-        self._u_cts = kwargs.get('u_countrates', [])
-        self._upper_limits = np.array(kwargs.get('upperlimits', []),
-                                      dtype=bool)
-        self._observation_types = np.array(kwargs['observation_types'])
-        self._observed = np.array(kwargs.get('observed', []), dtype=bool)
-
-        # Magnitudes first
-        # Note: Upper limits (censored data) currently treated as a
-        # half-Gaussian, this is very approximate and can be improved upon.
-        self._e_u_mags = [
-            kwargs['default_upper_limit_error']
-            if (e is None and eu is None and self._upper_limits[i]) else
-            (kwargs['default_no_error_bar_error']
-             if (e is None and eu is None) else (e if eu is None else eu))
-            for i, (e, eu) in enumerate(zip(self._e_mags, self._e_u_mags))
-        ]
-        self._e_l_mags = [
-            kwargs['default_upper_limit_error']
-            if (e is None and el is None and self._upper_limits[i]) else
-            (kwargs['default_no_error_bar_error']
-             if (e is None and el is None) else (e if el is None else el))
-            for i, (e, el) in enumerate(zip(self._e_mags, self._e_l_mags))
-        ]
-
-        # Now counts
-        self._cmask = self._observation_types == 'countrate'
-        self._cts[self._cmask] = -2.5 * np.log10(self._cts[self._cmask]
-                                                 .astype(np.float64))
-        self._e_u_cts = [
-            kwargs['default_upper_limit_error']
-            if (e is None and eu is None and self._upper_limits[i]) else
-            (kwargs['default_no_error_bar_error']
-             if (e is None and eu is None) else
-             2.5 * (np.log10(c + (e if eu is None else eu)) - np.log10(c)))
-            for i, (c, e, eu) in enumerate(zip(
-                self._cts, self._e_cts, self._e_u_cts))
-        ]
-        self._e_l_cts = [
-            kwargs['default_upper_limit_error']
-            if (e is None and el is None and self._upper_limits[i]) else
-            (kwargs['default_no_error_bar_error']
-             if (e is None and el is None) else
-             2.5 * (np.log10(c) - np.log10(c - (e if el is None else el))))
-            for i, (c, e, el) in enumerate(zip(
-                self._cts, self._e_cts, self._e_l_cts))
-        ]
-
-        # Now flux densities
-        self._e_u_fds = [
-            v if (e is None and eu is None and self._upper_limits[i]) else
-            (v if (e is None and eu is None) else (e if eu is None else eu))
-            for i, (e, eu, v) in enumerate(
-                zip(self._e_fds, self._e_u_fds, self._fds))
-        ]
-        self._e_l_fds = [
-            0.0 if self._upper_limits[i] else (
-                v if (e is None and el is None) else (e if el is None else el))
-            for i, (e, el, v) in enumerate(
-                zip(self._e_fds, self._e_l_fds, self._fds))
-        ]
-        self._fds = np.array([
-            x / flux_density_unit(y) if x is not None else None
-            for x, y in zip(self._fds, self._u_fds)
-        ])
-        self._e_u_fds = [
-            x / flux_density_unit(y) if x is not None else None
-            for x, y in zip(self._e_u_fds, self._u_fds)
-        ]
-        self._e_l_fds = [
-            x / flux_density_unit(y) if x is not None else None
-            for x, y in zip(self._e_l_fds, self._u_fds)
-        ]
